# api.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "User login",
        "description": "User logs into the application",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login Successfull",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - wrong email or password"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "get": {
        "operationId": "AuthController_logout",
        "summary": "User logout",
        "description": "User logs out of the application",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User logged out successfully"
          },
          "401": {
            "description": "Unauthorized request - invalid token"
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "User registration",
        "description": "Registers a new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Bad request - invalid email format"
          },
          "409": {
            "description": "Email already in use"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "get": {
        "operationId": "AuthController_refreshTokens",
        "summary": "Refresh access token",
        "description": "Refresh access token",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unauthorized - Refresh token expired or invalid"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/user/{user_id}/apartments": {
      "get": {
        "operationId": "ApartmentsController_getApartments",
        "summary": "List all apartments of a user",
        "description": "Returns all apartments owned by the specified user.",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Apartment list fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Apartment"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          }
        },
        "tags": [
          "apartments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ApartmentsController_create",
        "summary": "Create a new apartment",
        "description": "Creates a new apartment for the specified user.",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApartmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Apartment created successfully"
          },
          "400": {
            "description": "Bad Request – some fields are missing"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          }
        },
        "tags": [
          "apartments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user/{user_id}/apartments/{apartment_id}": {
      "get": {
        "operationId": "ApartmentsController_getById",
        "summary": "Get a single apartment",
        "description": "Fetches a specific apartment owned by the user.",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apartment_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Apartment fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Apartment"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          },
          "404": {
            "description": "Apartment not found"
          }
        },
        "tags": [
          "apartments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "ApartmentsController_update",
        "summary": "Update apartment details",
        "description": "Updates the details of an apartment.",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apartment_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Apartment updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          },
          "404": {
            "description": "Apartment not found"
          }
        },
        "tags": [
          "apartments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ApartmentsController_delete",
        "summary": "Delete an apartment",
        "description": "Deletes an apartment.",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apartment_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Apartment deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          },
          "404": {
            "description": "Apartment not found"
          }
        },
        "tags": [
          "apartments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "smart-home",
    "description": "Smart home application",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "example": "dl3@test.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "dl3test123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGci...",
            "description": "JWT Access Token"
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGci...",
            "description": "JWT Refresh Token"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "example": "dl3@test.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "dl3test123"
          },
          "firstName": {
            "type": "string",
            "description": "User first name, or any string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "avatarUrl": {
            "type": "string",
            "description": "URL for user avatar",
            "example": "http://example.com/avatar.png"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ]
      },
      "Apartment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "format": "uuid",
            "description": "Unique identifier for apartment"
          },
          "name": {
            "type": "string",
            "example": "My Home",
            "description": "Name of the apartment"
          },
          "location": {
            "type": "string",
            "example": "Oulu",
            "description": "Location of the apartment"
          }
        },
        "required": [
          "id",
          "name",
          "location"
        ]
      },
      "CreateApartmentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of an apartment",
            "example": "My Home"
          },
          "location": {
            "type": "string",
            "description": "City/area where the apartment is located",
            "example": "Oulu"
          }
        },
        "required": [
          "name",
          "location"
        ]
      },
      "UpdateApartmentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of an apartment",
            "example": "New Home"
          },
          "location": {
            "type": "string",
            "description": "City/area where the apartment is located",
            "example": "Helsinki"
          }
        }
      },
...
    }
  }
}