# apps/backend/src/apartments/dto/apartment-response-dto.ts
import { ApiProperty } from '@nestjs/swagger'
import { IsNotEmpty, IsString, IsUUID } from 'class-validator'

// Data Transfer Object (DTO) for apartment response
export class Apartment {
  @ApiProperty({
    example: '550e8400-e29b-41d4-a716-446655440000',
    format: 'uuid',
    description: 'Unique identifier for apartment',
  })
  @IsUUID()
  @IsNotEmpty()
  id: string

  @ApiProperty({ example: 'My Home', description: 'Name of the apartment' })
  @IsString()
  @IsNotEmpty()
  name: string

  @ApiProperty({ example: 'Oulu', description: 'Location of the apartment' })
  @IsString()
  @IsNotEmpty()
  location: string
}